bugfix: issue #83
