+++ pages/[lng]/login/index.tsx
@@ -2,11 +2,9 @@
 import { FC } from 'react'
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next'
 import { toast } from 'react-toastify'
-import dynamic from 'next/dynamic'
 import {
   Login,
   useI18n,
-  Widget,
   Logo
 } from '@sirclo/nexus'
 /* locales */
@@ -21,7 +19,6 @@ import useWindowSize from 'lib/useWindowSize'
 import Layout from 'components/Layout/Layout'
 import LoginRegisterOTP from 'components/LoginRegisterOTP'
 import Icon from 'components/Icon/Icon'
-const Placeholder = dynamic(() => import('components/Placeholder'))
 /* styles */
 import styles from 'public/scss/pages/Login.module.scss'
 
@@ -89,29 +86,13 @@ const LoginPage: FC<any> = ({
             onCompletedMsg={(msg: string) => toast.success(msg)}
             onErrorMsg={(msg: string) => toast.error(msg)}
             loadingComponent={
-              <Placeholder
-                classes={{
-                  placeholderList: styles.placeholderList,
-                }}
-                withList
-                listMany={3}
-              />
-              }
+              <div className="text-center d-flex">
+                <span className="spinner-border text-dark" role="status" />
+              </div>
+            }
             {...icons}
           />
         </section>
-        <Widget
-          widgetClassName={styles.widgetLogin}
-          pos="login-image"
-          loadingComponent={
-            <Placeholder
-              classes={{
-                placeholderImage: styles.widgetLogin
-              }}
-              withImage
-            />
-          }
-        />
       </LoginRegisterOTP>
     </Layout>
   )
