+++ public/scss/pages/Register.module.scss
@@ -82,19 +82,6 @@
   margin-bottom: 12px;
 }
 
-.widgetLogin
-{
-  @include position(fixed, 0, 0);
-  width: 50vw;
-  height: 100vh;
-  
-  img
-  {
-    width: 100% !important;
-  }
-}
-
-
 @media screen and (max-width: #{$breakpoint_max_md}) 
 {
   .section,
